// src/types/api.ts
export enum ApiErrorCode {
    VALIDATION_ERROR = "VALIDATION_ERROR",

    NOT_FOUND = "NOT_FOUND",

    PERMISSION_DENIED = "PERMISSION_DENIED",

    INVALID_TRANSITION = "INVALID_TRANSITION",

    DATABASE_ERROR = "DATABASE_ERROR",

    UNKNOWN_ERROR = "UNKNOWN_ERROR",

    AUTHENTICATION_ERROR = "AUTHENTICATION_ERROR",
       
}