// src/types/business.ts
export enum BusinessErrorCode {
    PRODUCT_NOT_FOUND = "PRODUCT_NOT_FOUND",

    REQUEST_NOT_FOUND_ERROR = "REQUEST_NOT_FOUND_ERROR",

    BRANCH_MISMATCH_ERROR = "BRANCH_MISMATCH_ERROR",

    LAST_LINE_DELETION_NOT_ALLOWED = "LAST_LINE_DELETION_NOT_ALLOWED",

    BANNED_REQUEST_TRANSITION = "BANNED_REQUEST_TRANSITION",

    NOT_A_WORKFLOW_TRANSITION = "NOT_A_WORKFLOW_TRANSITION",

    ZERO_QTY_PRODUCTION_NOT_ALLOWED = "ZERO_QTY_PRODUCTION_NOT_ALLOWED",

    ZERO_QTY_REQUEST_NOT_ALLOWED = "ZERO_QTY_REQUEST_NOT_ALLOWED",

    REQUEST_LINE_NOT_FOUND_ERROR = "REQUEST_LINE_NOT_FOUND_ERROR",

    ASSIGNMENT_MISMATCH_ERROR = "ASSIGNMENT_MISMATCH_ERROR",

    INSUFFICIENT_PAYLOAD = "INSUFFICIENT_PAYLOAD",

    FAULTY_BALANCE = "FAULTY_BALANCE",

    LINE_REQUEST_MISMATCH= "LINE_REQUEST_MISMATCH"
}